修复50ms扫描时间不准的问题

+++ Oscilloscope/layer/hal_tim.h
@@ -28,7 +28,7 @@ static int hal_tim_init(void);
 static void hal_test_init(void);
 void hal_pwm_stop(void);
 void hal_pwm_start(void);
-void hal_tim_psc(unsigned short psc);
+void hal_tim_psc(unsigned int psc);
 unsigned int hal_sys_time_us(void);
 unsigned int hal_tim2_cnt(void);
 unsigned int hal_tim3_cnt(void);
